t/rate-limiting/limitacion-de-transferencia. Also corrects small issues in other pages. fixes #36
