+++ public/index.html
@@ -71,7 +71,7 @@
       
         <li><a href="structs">Structs</a></li>
       
-        <li><a href="metodos">Metodos</a></li>
+        <li><a href="metodos">Métodos</a></li>
       
         <li><a href="interfaces">Interfaces</a></li>
       
@@ -97,15 +97,15 @@
       
         <li><a href="range-sobre-canales">Range sobre canales</a></li>
       
-        <li><a href="timers">Timers</a></li>
+        <li><a href="temporizadores">Temporizadores</a></li>
       
         <li><a href="tickers">Tickers</a></li>
       
         <li><a href="pool-de-trabajadores">Pool de trabajadores</a></li>
       
-        <li><a href="rate-limiting">Rate Limiting</a></li>
+        <li><a href="limitacion-de-transferencia">Limitación de transferencia</a></li>
       
-        <li><a href="atomic-counters">Atomic Counters</a></li>
+        <li><a href="contadores-atomicos">Contadores atómicos</a></li>
       
         <li><a href="mutexes">Mutexes</a></li>
