+++ public/metodos
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta http-eqiv="content-type" content="text/html;charset=utf-8">
-    <title>Go con Ejemplos: Metodos</title>
+    <title>Go con Ejemplos: Métodos</title>
     <link rel=stylesheet href="site.css">
   </head>
   <script>
@@ -17,7 +17,7 @@
   </script>
   <body>
     <div class="example" id="metodos">
-      <h2><a href="./">Go con Ejemplos</a>: Metodos</h2>
+      <h2><a href="./">Go con Ejemplos</a>: Métodos</h2>
       
       <table>
