+++ public/pool-de-trabajadores
@@ -207,7 +207,7 @@ hay 3 trabajadores operando de manera concurrente.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="rate-limiting">Rate Limiting</a>.
+        Siguiente ejemplo: <a href="limitacion-de-transferencia">Limitación de transferencia</a>.
       </p>
       
       <p class="footer">
