+++ public/range-sobre-canales
@@ -131,7 +131,7 @@ valores que no se han recibido.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="timers">Timers</a>.
+        Siguiente ejemplo: <a href="temporizadores">Temporizadores</a>.
       </p>
       
       <p class="footer">
