+++ public/structs
@@ -208,7 +208,7 @@ los apuntadores son automáticamente dereferenciados.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="metodos">Metodos</a>.
+        Siguiente ejemplo: <a href="metodos">Métodos</a>.
       </p>
       
       <p class="footer">
