+++ public/tickers
@@ -25,9 +25,9 @@
           <td class="docs">
             <p>Los <a>temporizadores</a> se usan cuando
 quieres hacer una cosa en el futuro - los <em>tickers</em>
-( por aquello de tic tac ) cuando cuando hacer algo
-repetidamente en intervalos regulares.
-Aquí hay un ejemmplo de un ticker que hace tic
+( por aquello de tic tac ) - se usan cuando se quiere
+hacer algo repetidamente en intervalos regulares.
+Aquí hay un ejemplo de un ticker que hace tic
 periodicamente hasta que lo detenemos.</p>
 
           </td>
@@ -42,7 +42,7 @@ periodicamente hasta que lo detenemos.</p>
             
           </td>
           <td class="code leading">
-          <a href="http://play.golang.org/p/YGlWOrly5k"><img title="Run code" src="play.png" class="run" /></a>
+          <a href="http://play.golang.org/p/frpKvXr0Ls"><img title="Run code" src="play.png" class="run" /></a>
             <div class="highlight"><pre><span class="kn">package</span> <span class="nx">main</span>
 </pre></div>
